Consolidate updating of reflective Field/Method references
Previously we used several different visitors to update Field &
Method references for structural redefinition. We also did not visit
or update JVMTI based references.

This consolidates all the visitors to a single
Runtime::VisitReflectiveTargets function with a single
ReflectiveTargetVisitor type. This simplifies the code around
structural redefinition and ensures that the reflective value holders
are in charge of the actual replacement.

Support was also added for walking internal openjdkjvmti references
for things like field-read/modification events.

Test: ./test.py --host
Bug: 134162467

Change-Id: Ic5fc1db7db0a30f947a1a67259dc024e149ebd57