Avoid using local refs in GetLocalVariableClosure
GetLocalVariableClosure would read an object value out of the stack,
then put it in a local-ref followed by a global-ref. This left an
opening for corruption if the closure was running for a suspended
thread and a GC was ongoing where the local table was updated
concurrently. By directly creating a global reference we should avoid
this issue.

Test: ./test.py --host
Test: ./test.py --host --all-compiler --all-gc -t 1966
Bug: 141590021

Change-Id: I5540c03bfdaecddbc1329ff6c7c58c6f18d6a090