Update patch set 1
Patch Set 1:

(I forgot to press submit on the comment)

What changes that we hit the DCHECK now? that was added quite a bit ago and we didn't have issues. I think it's a nice property to have everything verifiable in the boot image.

Patch-set: 1
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>