Always include `libart(d)-disassembler` in debug-able ART APEXes.
Library `libart(d)-disassembler` is required to execute some ART
run-tests (it's used to output disassembly information of compiled
methods to verify Checker `disassembly` checks); but it is dlopen'd by
`art::HGraphVisualizerDisassembler` (and not linked with
`libart(d)-compiler`). Make it an explicit dependency of all ART APEX
packages used for debugging or testing.

Test: art/build/apex/runtests.sh
Test: art/test/testrunner/testrunner.py --target --64 --ndebug --debug -t 1004-checker-volatile-ref-load
Bug: 140507091
Change-Id: I0ae206a26a873487d4415bb49b16e14718448ff5