Update patch set 13
Patch Set 13:

Hi Nicolas and Neeraj,

Very sorry for a delayed response!

In my previous comments I have tried to inquire on "Motivation and Feasability" points of the CL - e.g. benchmarking data - so you Google guys could have the data and evaluate it, the trade-offs and overhead. I believed checking on that first could be a good idea before reviewing the design, actual implementation and the code in detail (which can be done shortly after).

So probably to continue on that note:

I) As compile time for boot.oat is increased by ~1% then compile time impact on some bigger apps (like one social network app or similar) could be worth checking as I mentioned in my previous comments. I recall when working on my initial suberblock cloner patches I have seen quite unexpected impact on those big apps compile time.

II) Implementing and testing on other architectures. When I was pushing the initial change "ART: Enable scalar loop peeling and unrolling." (https://android-review.googlesource.com/c/platform/art/+/619387) it was suggested (and then implemented) that for such a high-level optimization it could be beneficial to enable it for all archs (unless there is a reason not to do so) to reduce the number of execution paths and for more robust testing. For instance in the current loop_analysis there are Default helper and target specific ones for fine-tuned heuristics. What do you think?

Thanks!

Patch-set: 13