Revert^2 "Walk internal ArtField/ArtMethod pointers"
This reverts commit 712fa800b2b78e527d36c88dc369bf4b723587ea.

We incorrectly didn't check if a method was obsolete before giving
its class's MethodIds array. This could lead to an (undefined)
arithmetic underflow, giving a calculated array offset outside the
acceptable range. To fix this we added back in the missing check.

Reason for revert: Fixed issue causing out-of-bounds array access
Test: ./test.py --host --debuggable --ndebuggable
Bug: 134162467

Change-Id: Iaffefeab6e889b4fb6554a11452d0af051001cb7