Update patch set 1
Patch Set 1:

> Patch Set 1: Presubmit-Ready+1
> 
> So today it works because we got lucky that all cases we use RecordImplicitNullCheck did get a known pc?

It is hard to say whether it is good luck or VIXL behaviour has not changed since the time ARM32 backend was written. It has TODOs regarding RecordImplicitNullCheck which state it would not happen in practice. The issue was discovered not because of some failures but whilst implementing some other functionality in ART.

>  If not, could we write a regression test?

I'll see whether it is possible to write a reliable test.

Patch-set: 1