Update patch set 1
Patch Set 1:

The problem here is that devices we run tests on may be on old platform images.

The changes to palette here should do something like probing for the new support and using it if found, or trying the new and falling back to the old otherwise.

Patch-set: 1
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>