Update patch set 1
Patch Set 1:

It seems slightly awkward that we need a lock to signal this state, but it has the advantage of fitting the sanitizer.

An alternative would be some evidence that a caller was within a a ScopedCodeCacheWrite block (e.g. a pointer to the scope, or some identifiable value from the SCCW instance), but not wild on this either.

Patch-set: 1
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>