Update patch set 2
Patch Set 2: Code-Review+1

> Patch Set 1:
> 
> The DotProduct for FP can get the rounding wrong. In an extreme case, you can get +Inf where consecutive calculation would get a finite number.

Hi Vladmir,
Thanks for the comment. In this case for the dot prod, we are emitting both multiplication and addition , so the rounding would happen twice as it would happen in the normal case. The rounding would have resulted in an error if the multiplication and addition were clubbed with a single instruction. I have added a test case for the floating point types.

Can you please give a test case where the rounding will result in an error?

Patch-set: 2
Reviewer: Gerrit User 1371444 <1371444@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1