Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review+1
> > > 
> > > Is this in preparation to take the actual fix?
> > 
> > No, this is the only one to take based on b/141519702#comment31 or should I take pa/1460265 instead?
> 
> I'm not sure how android10-gsi fits into the world, e.g. automergers, policy, etc. Just disabling is definitely safe and low risk.
> 
> If you were going to take the fix, then that could be cherry-picked from aosp too: https://r.android.com/1062191, but I wasn't aware of the cts failures before the bug here and have not tried the fix on cts.

The AOSP one can not be applied cleanly. android10-gsi is a leaf branch based on android10-release. I don't know enough what is going on here but I'm going to take this CL and assume it is a fix based on the comments in b/141519702. Let me know if there's a better way to do this.

Patch-set: 1
Reviewer: Gerrit User 1061057 <1061057@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0