Update patch set 2
Patch Set 2:

> Can you please give a test case where the rounding will result in an error?

The error happens because of the wrong order of addition.

Pick a finite number X such that 2.0 * X yields +Inf. Check the result of

    double[] a = new double[1024];
    for (int i = 0; i != 1024; ++i) a[i] = X;
    double[] b = new double[1024];
    for (int i = 0; i != 1024; ++i) b[i] = ((i & 1) == 0) ? 1.0 : -1.0;
    double result = testDotProdSimpleDouble(a, b);

For sequential addition, the `sum` jumps between 0 and X and the final result is 0. For the vectorized code, the accumulator will hold +Inf and -Inf from second iteration onward and I guess the final result will be NaN.

Patch-set: 2
Reviewer: Gerrit User 1018108 <1018108@85c56323-6fa9-3386-8a01-6480fb634889>