Ensure DexCaches stay in consistent state with structural redefinition.
In cases where we structurally redefined code where a use of a newly
added field or method dominates some instructions in a method the
process of redefinition will leave the new classes dex-cache in a
possibly inconsistent state. What happens is the verifier would see
the new method/field and decide that it was an unconditional throw
(since the class was loaded but it cannot find the ArtMethod/Field).

This would cause the verifier to not fully populate the DexCache.

Later during the reverify the methods would be put into the dex-cache
but the classes won't be (even if they can be found). Even if we
changed the verification code to manually place classes into the
dex-cache it would not solve this issue since if the class is not
already resolved we cannot do so with the mutator-lock we hold during
reverify.

This led to an inconsistent dex-cache where some methods were present
but the classes used to resolve them are not. In cases where the
interpreter attempts to perform access-checks it will fail to find the
class and crash.

To fix this we made the verifier skip the dead-code pruning on the
initial verification check so that classes will be resolved into the
new dex-cache. This means the verifier will resolve all classes
referred to while having a shared-mutator lock.

Test: ./test.py --host
Bug: 142876078

Change-Id: I0f1e8c47118cc84c8f23c4068944069ac74f5ea3