No longer collect the standard error in ART run-test outputs.
Recent Bionic linker changes now sometimes produce a warning on the
standard error, which affects the result checking phase of ART
run-tests when using Bionic (i.e. when running on an Android device or
on a Bionic-based host environment). Stop redirecting the standard
error to ART run-test output files to avoid this issue.

Note that we had already modified the ART run-tests to only print on
the standard output in order to produce deterministic test
results (printing messages on both the standard output and the
standard error would sometimes produce different results depending on
the execution). This change is a continuation of this policy.

Test: Run ART tests on device in a chroot environment
Bug: 143458513
Change-Id: I3e0a8f539e85e8b98cf50abd923ddfe10fb41dcf