Update patch set 1
Patch Set 1:

>   --- expected.txt	2018-08-24 19:09:07.282654807 +0100
>   +++ output.txt	2019-10-29 18:58:25.766337014 +0000
>   @@ -1,5 +1 @@
>   -Exception in thread "main" java.lang.BootstrapMethodError: Exception from call site #0 bootstrap method
>   -	at Main.main(Main.java:25)

Have you checked the history of the stderr redirects? This looks like it could be part of the test, to verify the error handling, so we might be losing test coverage if we drop it.

If so, we could perhaps introduce a special marker line that is printed directly before (and possibly also after) each runtest, and then only extract the text between the marker lines for comparison with the golden output.

Patch-set: 1
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>