Revert^2 "Package dexpreopt artifacts for libcore jars in the ART apex."
Reason: (some) build failures are fixed after reworking the patch.

This patch set adds precompiled dexpreopt files for the libcore part
of boot class path in the ART apex. The dexpreopt files are packaged
in javalib/$ARCH/ subdirectory and have names prefixed with 'boot'
to match the boot image on the system partition.

Test: m
Test: m com.android.art deapexer \
    && find $ANDROID_BUILD_TOP -type f -name 'com.android.art.*.apex \
        | xargs deapexer | grep boot \
    Expect to find javalib/$ARCH/boot*.{art,oat,vdex} files.
Test: art/build/apex/runtests.sh

Change-Id: I5a6e432a02f91fbfc7ed3771ced1fe73cfb46f04