Revert^4 "Package dexpreopt artifacts for libcore jars in the ART apex."
This reverts commit 0f7c792f12905558836cc01f3406d6ba364cefd7.

Reason for revert: coverage build with EMMA_INSTRUMENT_FRAMEWORK=true
is fixed by inspecting the environment variable and not generating
boot image in case it is set.

This patch set adds precompiled dexpreopt files for the libcore part
of boot class path in the ART apex. The dexpreopt files are packaged
in javalib/$ARCH/ subdirectory and have names prefixed with 'boot'
to match the boot image on the system partition.

Test: m
Test: m com.android.art deapexer \
    && find $ANDROID_BUILD_TOP -type f -name 'com.android.art.*.apex \
        | xargs deapexer | grep boot \
    Expect to find javalib/$ARCH/boot*.{art,oat,vdex} files.
Test: art/build/apex/runtests.sh

Bug: 144091989
Change-Id: I8378f42708a64339b0c9d70fd0d7804bc885859e