Update patch set 3
Patch Set 3:

Not sure if this is ready for review so ignore if still working on this. Currently, the testrunner.py script is still using subprocess.run() which was introduced in Python 3.5, but it appears we run in some environments that predate this.

Patch-set: 3
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>