Fix Core Platform API JNI check
Fix an inverted flag around kIsTargetBuild that broke check.

Simplify SharedObjectKind's so there are only values - part of the ART
module and other. Being part of the module is equivalent to
DexDomain::CorePlatform.

Add a test for Core Platform API checks for native code. The test is
derived from 674-hiddenapi, and separate because the existing test is
already complex.

Add 2030-core-platform-api-jni to the same knownfailures as
674-hiddenapi since the newer test likely to have the same failures.

Bug: 144502743
Test: art/test/run-test --host 2030
Test: art/test/run-test 2030
Test: erroneous warnings gone from master
Change-Id: I1a365a28b89bd46b3e403f6deb736f350c4baf10