Single-image boot image extensions.
Allow boot image extensions to be compiled into a single
art/oat/vdex triplet for any number of input dex files.
Support argument --image-fd so that this can be done with
file descriptors. This is intended for creating a boot
image extension in memory when starting zygote.

Remove the possibility to specify multiple oat or image
files. It was unused and untested.

Test: New test in dex2oat_image_test
Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Change-Id: I78e5410642a5df14c0891d87f27eefcb02156764