Update patch set 1
Patch Set 1: Code-Review+2

I was a bit concerned whether we could do this more efficiently in software by using integer comparisons. But the (not well-documented?) intent here seems to be that NaNs should compare equal, which means you need a NaN test on the common paths. So I'm inclined to believe this is the best approach.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2