Make opaque-jni-ids:swapable more efficient
-Xopaque-jni-ids:swapable was rather memory inefficient, allocating an
entire PointerSizedArray to mark a method/field as having a
Pointer-id. Since normally these will not be used once the JNI id type
moves to kPointer this is rather wasteful. By changing to instead have
a single marker object which is used instead of the full PointerArray
we can use significantly less memory in common use-cases.

Test: Manual
Test: count-fields.py
Bug: 134162467
Change-Id: I224f0810373afef2335fad092eb758f91455f9e5