Update patch set 3
Patch Set 3: Presubmit-Ready+1

This looks kind of like it might be a real failure. But it seems to be timing out with all but the main thread in Object.Wait(), and presumably suspended, while executing

  STLDeleteElements(&continuous_spaces_);

in the Heap destructor on the main thread. The main thread appears to be runnable and running when it gets a signal. (See https://android-build.googleplex.com/builds/pending/P10731134/art-test/latest/view/logs/build.log, search for timeout.)

b/121302864 could conceivably be related, but probably not.

I cannot easily reproduce this locally. I'll retry Treehugger, mostly to see whether this fails in exactly the same place.

Patch-set: 3
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1