Use apex_available property
Use apex_available property to prevent modules that are only for the ART
apex from being used outside of the APEX.

Some modules are also marked as "//apex_available:platform_noinstall"
because the platform variant of the modules are still needed; by some
modules in /system/core and by some ART tests. Note that the modules are
still prevented from being installed to the system partition, unless
they are marked as "//apex_available:platform".

Bug: 128708192
Bug: 133140750
Test: m
Change-Id: I990e0a67e066d9b1dc221748422a2f44449258da