RFC: Implement basic predicated SIMD vectorization.
NOT YET FOR SUBMISSION, depends on the VIXL update.

This CL brings support for predicated execution for
auto-vectorizer and implements arm64 SVE vector backend.

This version passes all the VIXL simulator-runnable tests in
SVE mode with checker off (as all VecOp CHECKs need to be
adjusted for an extra input) and all tests in NEON mode.

Test: art SIMD tests on VIXL simulator.
Test: test-art-target on Arm fvp

Change-Id: Ib78bde31a15e6713d875d6668ad4458f5519605f