Update patch set 5
Patch Set 5:

(9 comments)

Sorry for the delay, I had a partial review but not ready.

The code if the oat file setup is .... pretty hard to review (not the fault of this CL :) )

I think this should be ok, my main comments are about adding more docs in the code with examples.

There are some tests you should validate in master:
-> do we correctly pick things up after adb install?
-> do we correctly record the path in the profile. you can use profman --dump-only and check the dex file names.
-> for dynamic code (secondary dex files), pick gms for example, do we pick the oat files correctly? Note that prebuilds and updates behave differently due to symlinks.

Patch-set: 5
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>