Update patch set 1
Patch Set 1:

Should I make this change for libartbase, libprofile, and libdexfile? I think libartbase and libdexfile should be modified the same way, but I couldn't find any bugs where libprofile.so was in the stack trace, and it's not clear if this is something that you expect to see in stack traces. If it is, then debug frame should be saved for that library too.

Patch-set: 1
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>