Update patch set 2
Patch Set 2:

ngeoffray, It seems that the failure was unrelated. I'm going to add in more logging to try to narrow down where the flake is.

b/146436130 is tracking this effort.

For now I don't think there's any reason not to push this back in.

Patch-set: 2
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>